Выполнила модуляцию
